Fix docs build on RtD (need to mock swig-generated code)
The bug was introduced in e3b590de.

If `tabu.tabu_search` file is not properly mocked while building docs
with:
```
autodoc_mock_imports = ["tabu.tabu_search"]
```
`TabuSampler` import fails (and the docstring is not compiled).

The reason we need to mock it is because otherwise we would need to
build it. And to build it, we need `swig`, which is not available on RtD
(see #25). The only alternative IIRC is to roll out our docker build
container.